+++ .eslintrc.json
@@ -34,6 +34,7 @@
     "jsx-a11y/media-has-caption": "off",
     "no-nested-ternary": "off",
     "no-console": "off",
+    "@typescript-eslint/no-unused-vars": "off",
     "no-alert": "off",
     "react-hooks/exhaustive-deps": "off",
     "react/button-has-type": "off",
